+++ tests/unit/domain/user/test_user_entity.py
@@ -1,5 +1,6 @@
 import pytest
 
+from domain.auth.auth_info.value_objects import HashedPassword
 from domain.user.exceptions import (
     PasswordMissingLowercaseError,
     PasswordMissingNumberError,
@@ -9,7 +10,7 @@ from domain.user.exceptions import (
     PasswordTooShortError,
 )
 from domain.user.user import User
-from domain.user.value_objects import Email, HashedPassword, PlainPassword, Username
+from domain.user.value_objects import Email, PlainPassword, Username
 from tests.unit.conftest import FakeHasher, FakeTimeProvider
